Use int for audio devices' user priority consistently
Bug: b:251698397
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3952774
Commit-Queue: Li-Yu Yu <aaronyu@google.com>
Cr-Commit-Position: refs/heads/main@{#1059596}

+++ chromeos/ash/components/audio/audio_devices_pref_handler_impl_unittest.cc
@@ -179,7 +179,7 @@ class AudioDevicesPrefHandlerTest : public testing::TestWithParam<bool> {
                          : audio_pref_handler_->GetOutputVolumeValue(&device);
   }
 
-  double GetUserPriority(const AudioDevice& device) {
+  int GetUserPriority(const AudioDevice& device) {
     return audio_pref_handler_->GetUserPriority(device);
   }
